+++ mobile/android/android-components/docs/changelog.md
@@ -4,13 +4,21 @@ title: Changelog
 permalink: /changelog/
 ---
 
-# 41.0.0-SNAPSHOT (In Development)
+# 42.0.0-SNAPSHOT (In Development)
 
-* [Commits](https://github.com/mozilla-mobile/android-components/compare/v40.0.0...master)
-* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/101?closed=1)
+* [Commits](https://github.com/mozilla-mobile/android-components/compare/v41.0.0...master)
+* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/102?closed=1)
 * [Dependencies](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Dependencies.kt)
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
+
+# 41.0.0
+
+* [Commits](https://github.com/mozilla-mobile/android-components/compare/v40.0.0...v41.0.0)
+* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/101?closed=1)
+* [Dependencies](https://github.com/mozilla-mobile/android-components/blob/v41.0.0/buildSrc/src/main/java/Dependencies.kt)
+* [Gecko](https://github.com/mozilla-mobile/android-components/blob/v41.0.0/buildSrc/src/main/java/Gecko.kt)
+* [Configuration](https://github.com/mozilla-mobile/android-components/blob/v41.0.0/buildSrc/src/main/java/Config.kt)
 
 * **feature-tabs**
   * Fixed issue [#6907](https://github.com/mozilla-mobile/android-components/issues/6907). Uses MediaState when mapping BrowserState to tabs.